+++ package.py
@@ -1,24 +1,24 @@
 # pylint: disable=invalid-name
 """Shotgun_api3"""
-name = 'shotgun_api3'
+name = "shotgun_api3"
 
-_shotgunSoftwareVersion = '3.1.1'
-_rdoVersion = '1.1.1'
-version = '{0}-rdo-{1}'.format(_shotgunSoftwareVersion, _rdoVersion)
+_shotgunSoftwareVersion = "3.3.4"
+_rdoVersion = "1.0.0"
+version = "{0}-rdo-{1}".format(_shotgunSoftwareVersion, _rdoVersion)
 
-authors = ['shotgundev@rodeofx.com']
+authors = ["shotgundev@rodeofx.com"]
 
-description = 'Fork of the python api of shotgun.'
+description = "Fork of the python api of shotgun."
 
-requires = ['python-2.6|2.7|3.7+']
+requires = ["python-2.7|3.7+"]
 
-private_build_requires = ['rdo_package_utils']
+private_build_requires = ["rdo_package_utils"]
 
-build_command = 'python {root}/build.py {install}'
+build_command = "python {root}/build.py {install}"
 
-uuid = '9E411E66-9F35-49BC-AC2E-E9DC6D50D109'
+uuid = "9E411E66-9F35-49BC-AC2E-E9DC6D50D109"
 
 
 def commands():
     """Commands"""
-    env.PYTHONPATH.append('{root}/')
+    env.PYTHONPATH.append("{root}/")