+++ shotgun_api3/lib/mockgun/mockgun.py
@@ -193,13 +193,15 @@ class Shotgun(object):
         # they way they would expect to in the real API.
         self.config = _Config(self)
 
+        self.config.set_server_params(base_url)
+
         # load in the shotgun schema to associate with this Shotgun
         (schema_path, schema_entity_path) = self.get_schema_paths()
 
         if schema_path is None or schema_entity_path is None:
             raise MockgunError("Cannot create Mockgun instance because no schema files have been defined. "
                                "Before creating a Mockgun instance, please call Mockgun.set_schema_paths() "
-                               "in order to specify which Shotgun schema Mockgun should operate against.")
+                               "in order to specify which ShotGrid schema Mockgun should operate against.")
 
         self._schema, self._schema_entity = SchemaFactory.get_schemas(schema_path, schema_entity_path)
 
@@ -504,7 +506,7 @@ class Shotgun(object):
                                    "entity_type": six.string_types}[sg_type]
                 except KeyError:
                     raise ShotgunError(
-                        "Field %s.%s: Handling for Shotgun type %s is not implemented" %
+                        "Field %s.%s: Handling for ShotGrid type %s is not implemented" %
                         (entity_type, field, sg_type)
                     )
 
@@ -842,8 +844,6 @@ class Shotgun(object):
             field_type = self._get_field_type(entity_type, field)
             if field_type == "entity" and data[field]:
                 row[field] = {"type": data[field]["type"], "id": data[field]["id"]}
-            elif field_type == "multi_entity":
-                row[field] = [{"type": item["type"], "id": item["id"]} for item in data[field]]
             else:
                 row[field] = data[field]